+++ packages/client/src/index.ts
@@ -29,6 +29,7 @@ export {
   usePageParams,
   usePageSearch,
 } from "./framework/page/page-context.js";
+export { useRouteParams } from "./framework/page/route-context.js";
 export {
   getFnQueryKey,
   getFnQueryOptions,
@@ -47,6 +48,7 @@ export type { AppRouteContext } from "./standalone/context.js";
 export { createAppRootRoute } from "./standalone/context.js";
 export type {
   ActiveLinkOptions,
+  HrefResolver,
   LinkOptions,
   LinkProps,
   NavigateOptions,
@@ -62,6 +64,8 @@ export {
   Navigate,
   notFound,
   redirect,
+  useHref,
+  useHrefResolver,
   useLinkProps,
   useLocation,
   useNavigate,